+++ src/config/theme/mod.rs
@@ -9,6 +9,7 @@ use ratatui::style::{Color, Style};
 
 use self::{
     header::{HeaderConfig, HeaderConfigFile},
+    lyrics::{LyricsConfig, LyricsConfigFile},
     progress_bar::{ProgressBarConfig, ProgressBarConfigFile},
     queue_table::{QueueTableColumns, QueueTableColumnsFile},
     scrollbar::ScrollbarConfig,
@@ -18,6 +19,7 @@ use crate::mpd::commands::metadata_tag::MetadataTag;
 
 mod header;
 pub mod level_styles;
+mod lyrics;
 mod progress_bar;
 pub mod properties;
 mod queue_table;
@@ -68,6 +70,7 @@ pub struct UiConfig {
     pub format_tag_separator: String,
     pub mutliple_tag_resolution_strategy: TagResolutionStrategy,
     pub level_styles: LevelStyles,
+    pub lyrics: LyricsConfig,
 }
 
 #[derive(Debug, Serialize, Deserialize, PartialEq, Eq)]
@@ -111,6 +114,8 @@ pub struct UiConfigFile {
     pub(super) mutliple_tag_resolution_strategy: TagResolutionStrategy,
     #[serde(default)]
     pub(super) level_styles: LevelStylesFile,
+    #[serde(default)]
+    pub(super) lyrics: LyricsConfigFile,
 }
 
 impl Default for UiConfigFile {
@@ -182,6 +187,7 @@ impl Default for UiConfigFile {
             },
             level_styles: LevelStylesFile::default(),
             components: HashMap::default(),
+            lyrics: LyricsConfigFile::default(),
         }
     }
 }
@@ -379,6 +385,7 @@ impl TryFrom<UiConfigFile> for UiConfig {
                 .preview_metadata_group_style
                 .to_config_or(None, None)?,
             level_styles: value.level_styles.try_into()?,
+            lyrics: value.lyrics.into(),
         })
     }
 }