+++ CHANGELOG.md
@@ -53,6 +53,7 @@ lists they are listed as multiple entries.
 - Added `maps_three_symbols` test for progress bar. This will help avoid any errors while changing progress bar code in future
 - Added `PopConfigErrorModal` so theat the config error modals are automatically removed when the config reloads and is found correct
 - Added style configuration for dir and song symbols in browsers
+- Added `lyrics` config option `timestamp` for showing line timestamp
 
 ### Changed